Separates the route
