+++ api/routes/rsvp.js
@@ -4,3 +4,5 @@ import rsvp from '../controllers/rsvp';
 const routes = Router();
 
 routes.post('/', rsvp.respondToRsvp);
+
+export default routes;
