+++ api/server.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import routes from './routes/index';
+import routes from './routes';
 
 const app = express();
 
@@ -7,7 +7,7 @@ const port = process.env.PORT || 8080;
 
 app.use(express.json());
 
-app.use('/', routes);
+app.use('/api/v1', routes);
 
 app.listen(port, () => {
   console.log(`Sever running on port ${port}`);
