+++ package.json
@@ -33,6 +33,7 @@
     "eslint": "^5.11.1",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-plugin-import": "^2.14.0",
+    "joi": "^14.3.1",
     "mocha-lcov-reporter": "^1.3.0"
   }
 }
